moved test requirements outside module
