+++ .travis.yml
@@ -1,7 +1,6 @@
 language: python
 python:
   - "2.7"
-  - "pypy"
 install:
   - sudo pip install -r test_requirements.txt 
 before_script:
