+++ test_requirements.txt
@@ -3,3 +3,4 @@ fudge==1.0.3
 pep8==1.4.6
 pylint==1.0.0
 nose
+docker
